cerbero: Don't try to remove non-existent downloads
Same code exists for wget and urllib2.

+++ cerbero/utils/shell.py
@@ -332,7 +332,8 @@ def download_curl(url, destination=None, recursive=False, check_cert=True, overw
         try:
             call(cmd, path)
         except FatalError, e:
-            os.remove(destination)
+            if os.path.exists(destination):
+                os.remove(destination)
             raise e
